requirements.txt
pyodbc
pandas
scikit-learn
python-dotenv
